nvme: Move tr_send to pci.c
For every nvme device there can be several cdqs. They are identified
with an offset in the cdq array member. This is the id used by the user.
Given that the user passes an array offset, we need to get the tr_send
data from within pci.c as the array lives there.

Move the logic and correct the confusion between the kernel id and the
controller id.

Signed-off-by: Joel Granados <joel.granados@kernel.org>