block: allow surpressing automatic integrity metadata generation
Some file systems have checksums and self-describing metadata that
provide a protection enveloped that is as good as or better as that
provided by common protection information formats.  For metadata which
is small and synchronously, read the overhead of the automatic
integrity data generation in the block layer can be noticable, so
add a flag to surpress it when the format allows automatic insertation
and stripping instead.  The flag is simply ignored when the device does
not support integrity metadata, or when the device can't strip and insert
it.

Signed-off-by: Christoph Hellwig <hch@lst.de>