nvme/cdq: Remove all code to traverse CDQ
With the CDQ memory being mmaped into user space. The responsibility of
making sense of that memory is now on user space. The kernel remains
responsible of signaling an eventfd when it detects that a ONE_SHOT CDQ
AEN.

Interface is simplified as there is no need to define the number of
entries; we can get away with just the total CDQ size in number of
bytes. There is only need for the MOS and the CDS for CDQ creation, the
rest is removed. We are left with six values at CDQ creation: size of
the CDQ in bytes, eventfd for the TPT, the returned cdq identifier and
the internal kernel file descriptor, CQS and MOS.

Signed-off-by: Joel Granados <joel.granados@kernel.org>